requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
beautifulsoup4==4.12.2
# Use lighter alternatives for Render deployment
# opencv-python-headless==4.8.1.78
# deepface==0.0.79 